tests/submission/Infraware/Forms/contents/Forms/input_formnovalidate.html

<!DOCTYPE HTML>
<html>
 <head>
  <title>Forms</title>
  <script src="../../resources/testharness.js"></script>
  <script src="../../resources/testharnessreport.js"></script>
 </head>
 <body>
	<p>
		<h3>input_formNoValidate</h3>
	</p>
	
	<hr>

	<div id="log"></div> 

	<form method="post"
		  enctype="application/x-www-form-urlencoded"
		  action=""
		  name="input_form">
	<p><input type='hidden' id='input_text' required></p>
	</form>

	<script>
		
		var input_text = document.getElementById("input_text");
		input_text.formNoValidate = true;

		if (typeof(input_text.formNoValidate) == "boolean") {
			test(function() {
				assert_equals(input_text.formNoValidate, true, "formNoValidate attribute is not correct.");
			});
		} else {
			test(function() {
				assert_unreached("formNoValidate attribute is not exist.");
			});
		}

	</script>

 </body>
</html>
